refactor(compaction): split CompactionService into runner + queue (max-params 2/3)
CompactionService injected 7 deps (surreal, apiKeys, config, metrics,
SUMMARY_GENERATOR, claim, workerLoop). Three-way responsibility split:
- CompactionRunnerService (surreal, config, generator) — the retention
  engine. compactCompany(companyId) and compactAll(tenants[]) now take
  the tenant list as a parameter, so it carries no apiKeys dep and
  returns stats metrics-free.
- CompactionQueueService (apiKeys, claim, workerLoop) — handler
  registration, per-tenant cron enqueue, queue-mode flag, tenant list.
- CompactionService (runner, queue, metrics) — the cron/dispatch shell;
  emits per-tenant metrics and keeps the public compactAll/compactCompany
  entry points the admin + dreams paths call.

SUMMARY_GENERATOR token + CompactionStats moved to compaction.types.ts
(re-exported from compaction.service for backward compat). Each class is
now ≤3 deps.

Part of the max-params=3 god-class split program.
